provide buffer-local toggling of electric insert
Allow the user to toggle electric insert behavior on/off via a
'realign_electric_dot' buffer-local variable. The global variable
'realign_ruby_electric_dot' controls whether the mapping is set up at
all, and the buffer local 'realign_electric_dot' is evaluated each time
the mapping is executed, and decides whether to apply the realign
operation.

I need to find a way to refactor this so it doesn't require a
source-code line of 144 characters!